refactor: move Health sub-components into views/components
